[YS-513] refactor: 컴포넌트 분리하여 스타일링 의존성 제거
