+++ src/app/user/profile/mobile/edit/[infoType]/components/EditFormLayout/EditFormLayout.css.ts
@@ -1,0 +1,9 @@
+import { style } from '@vanilla-extract/css';
+
+export const mainContentLayout = style({
+  display: 'flex',
+  flexDirection: 'column',
+  gap: '2rem',
+  padding: '2.8rem 1.6rem 0.2rem 1.6rem',
+  overflowY: 'auto',
+});
