+++ src/app/user/profile/mobile/edit/[infoType]/components/MatchTypeEditForm.tsx
@@ -2,7 +2,7 @@
 
 import { FormProvider } from 'react-hook-form';
 
-import EditFormLayout from './EditFormLayout';
+import EditFormLayout from './EditFormLayout/EditFormLayout';
 
 import { ParticipantResponse } from '@/apis/login';
 import { MatchType } from '@/app/join/JoinPage.types';
