+++ src/app/user/profile/mobile/edit/[infoType]/components/NameEditForm.tsx
@@ -2,7 +2,7 @@
 
 import { FormProvider } from 'react-hook-form';
 
-import EditFormLayout from './EditFormLayout';
+import EditFormLayout from './EditFormLayout/EditFormLayout';
 
 import { ParticipantResponse } from '@/apis/login';
 import JoinInput from '@/app/join/components/JoinInput/JoinInput';
